move fixtures to tests/integration/fixtures.py
