+++ tests/integration/fixtures.py
@@ -1,0 +1,65 @@
+import os
+import subprocess
+
+import py.path
+import pytest
+
+
+EXECUTABLE = os.getenv(
+    "MAPILLARY_TOOLS_EXECUTABLE", "python3 -m mapillary_tools.commands"
+)
+IMPORT_PATH = "tests/integration/mapillary_tools_process_images_provider/data"
+USERNAME = "test_username_MAKE_SURE_IT_IS_UNIQUE_AND_LONG_AND_BORING"
+
+
+@pytest.fixture
+def setup_config(tmpdir: py.path.local):
+    config_path = tmpdir.mkdir("configs").join("CLIENT_ID")
+    os.environ["MAPILLARY_CONFIG_PATH"] = str(config_path)
+    x = subprocess.run(
+        f"{EXECUTABLE} authenticate --user_name {USERNAME} --jwt test_user_token",
+        shell=True,
+    )
+    assert x.returncode == 0, x.stderr
+    yield config_path
+    if tmpdir.check():
+        tmpdir.remove(ignore_errors=True)
+    del os.environ["MAPILLARY_CONFIG_PATH"]
+
+
+@pytest.fixture
+def setup_data(tmpdir: py.path.local):
+    data_path = tmpdir.mkdir("data")
+    source = py.path.local(IMPORT_PATH)
+    source.copy(data_path)
+    yield data_path
+    if tmpdir.check():
+        tmpdir.remove(ignore_errors=True)
+
+
+@pytest.fixture
+def setup_upload(tmpdir: py.path.local):
+    upload_dir = tmpdir.mkdir("mapillary_public_uploads")
+    os.environ["MAPILLARY_UPLOAD_PATH"] = str(upload_dir)
+    os.environ["MAPILLARY__DISABLE_BLACKVUE_CHECK"] = "YES"
+    os.environ["MAPILLARY__DISABLE_CAMM_CHECK"] = "YES"
+    yield upload_dir
+    if tmpdir.check():
+        tmpdir.remove(ignore_errors=True)
+    del os.environ["MAPILLARY_UPLOAD_PATH"]
+    del os.environ["MAPILLARY__DISABLE_BLACKVUE_CHECK"]
+
+
+def ffmpeg_installed():
+    ffmpeg_path = os.getenv("MAPILLARY_TOOLS_FFMPEG_PATH", "ffmpeg")
+    ffprobe_path = os.getenv("MAPILLARY_TOOLS_FFPROBE_PATH", "ffprobe")
+    try:
+        subprocess.run([ffmpeg_path, "-version"])
+        # In Windows, ffmpeg is installed but ffprobe is not?
+        subprocess.run([ffprobe_path, "-version"])
+    except FileNotFoundError:
+        return False
+    return True
+
+
+is_ffmpeg_installed = ffmpeg_installed()
