+++ tests/integration/test_history.py
@@ -2,10 +2,9 @@ import os
 import subprocess
 
 import py.path
-
 import pytest
 
-from .test_process import EXECUTABLE, setup_config, setup_data, setup_upload, USERNAME
+from .fixtures import EXECUTABLE, setup_config, setup_data, setup_upload, USERNAME
 
 
 @pytest.fixture
@@ -20,10 +19,10 @@ def setup_history_path(tmpdir: py.path.local):
     del os.environ["MAPILLARY__ENABLE_UPLOAD_HISTORY_FOR_DRY_RUN"]
 
 
+@pytest.mark.usefixtures("setup_config")
+@pytest.mark.usefixtures("setup_history_path")
 def test_upload_images(
     setup_data: py.path.local,
-    setup_config: py.path.local,
-    setup_history_path: py.path.local,
     setup_upload: py.path.local,
 ):
     assert len(setup_upload.listdir()) == 0
@@ -47,10 +46,10 @@ def test_upload_images(
     ), "should NOT upload because it is uploaded already"
 
 
+@pytest.mark.usefixtures("setup_config")
+@pytest.mark.usefixtures("setup_history_path")
 def test_upload_blackvue(
     setup_data: py.path.local,
-    setup_config: py.path.local,
-    setup_history_path: py.path.local,
     setup_upload: py.path.local,
 ):
     assert len(setup_upload.listdir()) == 0
